@@ -3,6 +3,6 @@ fn main() {
     let mut s1 = String::from("foo");
     let s2 = "bar";
     s1.push_str(s2);
-    println!("s2 is {}", s2);
+    println!("s2 is {s2}");
     // ANCHOR_END: here
 }
