@@ -4,6 +4,6 @@ fn main() {
     let s2 = String::from("tac");
     let s3 = String::from("toe");
 
-    let s = format!("{}-{}-{}", s1, s2, s3);
+    let s = format!("{s1}-{s2}-{s3}");
     // ANCHOR_END: here
 }
