@@ -10,18 +10,16 @@ fn main() {
 
     match msg {
         Message::Quit => {
-            println!("The Quit variant has no data to destructure.")
+            println!("The Quit variant has no data to destructure.");
         }
         Message::Move { x, y } => {
-            println!(
-                "Move in the x direction {} and in the y direction {}",
-                x, y
-            );
+            println!("Move in the x direction {x} and in the y direction {y}");
         }
-        Message::Write(text) => println!("Text message: {}", text),
-        Message::ChangeColor(r, g, b) => println!(
-            "Change the color to red {}, green {}, and blue {}",
-            r, g, b
-        ),
+        Message::Write(text) => {
+            println!("Text message: {text}");
+        }
+        Message::ChangeColor(r, g, b) => {
+            println!("Change the color to red {r}, green {g}, and blue {b}",)
+        }
     }
 }
