@@ -8,6 +8,6 @@ fn main() {
     scores.insert(String::from("Yellow"), 50);
 
     let team_name = String::from("Blue");
-    let score = scores.get(&team_name);
+    let score = scores.get(&team_name).copied().unwrap_or(0);
     // ANCHOR_END: here
 }
