@@ -1,33 +1,10 @@
 use std::thread;
-use std::time::Duration;
-
-// ANCHOR: here
-fn generate_workout(intensity: u32, random_number: u32) {
-    let expensive_closure = |num| {
-        println!("calculating slowly...");
-        thread::sleep(Duration::from_secs(2));
-        num
-    };
-
-    if intensity < 25 {
-        println!("Today, do {} pushups!", expensive_closure(intensity));
-        println!("Next, do {} situps!", expensive_closure(intensity));
-    } else {
-        if random_number == 3 {
-            println!("Take a break today! Remember to stay hydrated!");
-        } else {
-            println!(
-                "Today, run for {} minutes!",
-                expensive_closure(intensity)
-            );
-        }
-    }
-}
-// ANCHOR_END: here
 
 fn main() {
-    let simulated_user_specified_value = 10;
-    let simulated_random_number = 7;
+    let list = vec![1, 2, 3];
+    println!("Before defining closure: {:?}", list);
 
-    generate_workout(simulated_user_specified_value, simulated_random_number);
+    thread::spawn(move || println!("From thread: {:?}", list))
+        .join()
+        .unwrap();
 }
