@@ -2,8 +2,5 @@ use add_one;
 
 fn main() {
     let num = 10;
-    println!(
-        "Hello, world! {num} plus one is {}!",
-        add_one::add_one(num)
-    );
+    println!("Hello, world! {num} plus one is {}!", add_one::add_one(num));
 }
